edge: move peerconnection shim to separate file
moves the peerconnection shim to a separate file
